updated document, moved version extract function to version utils file
